Set simulator output to data folder. Re-organized modules
