+++ .gitignore
@@ -37,4 +37,3 @@ bcipy/language/out/
 
 
 bcipy/simulator/tests/resource/
-bcipy/simulator/generated/SIM_*
